PP-1409 addressing review comments
 - using FIELD_ constants to access auth request params
 - Refactored resource tests to multiple classes
 - Changed 423 status code to 401 with an error message.
